+++ test/clientlib/stateful_module_dep_rules_test.exs
@@ -31,14 +31,6 @@ defmodule Argus.Clientlib.StatefulModuleDepRulesTest do
       rules = """
       .include "#{Path.join(priv_dl(), "clientlib/imports.dl")}"
 
-      .decl implements_behaviour(mod: symbol, behaviour: symbol)
-      .input implements_behaviour
-
-      .decl sync_call(caller_func: symbol, callee_mod: symbol)
-      .input sync_call
-
-      .decl async_cast(caller_func: symbol, callee_mod: symbol)
-      .input async_cast
 
       .include "#{Path.join(priv_dl(), "clientlib/otp.dl")}"